refactor: rename old effect file
