+++ cogs5e/initiative/__init__.py
@@ -1,6 +1,6 @@
 from .combat import Combat
 from .combatant import Combatant, MonsterCombatant, PlayerCombatant
-from .effect import Effect
+from .effect_old import Effect
 from .errors import *
 from .group import CombatantGroup
 from .types import CombatantType
